// app/src/core/createFbPixels/DTO.ts
export interface CreateFbPixelsDTO_I {
  name: string;
  accountId: number;
  pixel_id: string; // 2170386460137024
  access_token: string; // EAAIgMgx7yggBOZBePnmRhm8LZCiaX8qyqWK9flSlW22uS3NZBeZC0SRuWAwdUBT9hrZC4XOAl6NWF04Sb5pQg2FEYl8Jj967e1nNMzZA5QyqlPCWO7e6zKuI67Ghvqtuk25jIYBIZCYuXR9FF3UnPIe6LDFFuSMziEKZBuHZAKXuZCx8n6FHDbi5QscZBtct48ZARBtozAZDZD
  status?: boolean;
  businessId?: number;
}